extract runtime pallets part.6
