+++ runtime/crab/src/pallets/relay_authorities.rs
@@ -10,7 +10,7 @@ use darwinia_relay_primitives::relay_authorities::OpCode;
 
 frame_support::parameter_types! {
 	pub const EthereumRelayAuthoritiesLockId: LockIdentifier = *b"ethrauth";
-	pub const EthereumRelayAuthoritiesTermDuration: BlockNumber = 30 * DAYS;
+	pub const EthereumRelayAuthoritiesTermDuration: BlockNumber = 7 * DAYS;
 	pub const MaxCandidates: usize = 7;
 	pub const OpCodes: (OpCode, OpCode) = (
 		[71, 159, 189, 249],
